requirements.txt
pyinstaller 
argcomplete 
click 
colorama 
GitPython 
PyGithub